requirements.txt
# Gmail Auto-responder Requirements

# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.23.2
jinja2==3.1.2
python-multipart==0.0.6
starlette==0.27.0

# Google API
google-api-python-client==2.108.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1

# OpenAI
openai==1.3.5

# Scheduler
apscheduler==3.10.4

# Database
sqlalchemy==2.0.23

# Environment Variables
python-dotenv==1.0.0

# Email Handling
email-validator==2.1.0
